Added All Images from Course Manual odt
We will be adding all the images by link to this folder to the adoc file.
I went to the Course Manual.odt and right clicked on each image and saved it into the images folder.
Next step will be to update the adoc file to point to these images.
This will have the advantage that if in the future we need to update an image, we can update it and as long as we keep the same name, the adoc will point to the correct image.