Updated images inside course manual adoc
Renamed images to have the name of the document cm_ followed by the picture
name numbered.

I added the numbers for easier editing of the document. The numbers are of the
format 00X.00 so, if in the future we need to add an image in between two numbers,
we can use the decimals to still organize them in order of appearance in the directoy without having to rename all the rest of the images.

I couldn't find the imagesdir property of the documents, so I am assuming that is
something Oliver set via Antora.